Pull biological data from the PacFIN database

Pull biological data from the Comprehensive Biological Data (comprehensive_bds_comm) table in the PacFIN database (PACFIN_MARTS). Data are in their raw form and should be cleaned either by hand or using cleanPacFIN.

PullBDS.PacFIN(
  pacfin_species_code,
  username = getUserName("PacFIN"),
  password,
  savedir = getwd(),
  verbose = TRUE
)

Arguments

pacfin_species_code

A character string for the PacFIN species code that you desire to pull data for. Known also as "SPID" in the PacFIN database if using legacy sql scripts. An example for sablefish would be pacfin_species_code = "SABL". A list of pacfin_species_code can be found online: https://pacfin.psmfc.org/pacfin_pub/marketcat.php

username

The username you would like to use to access data from datasourcename. The default is to generate username from your computer name.

password

Your password for the datasourcename.

savedir

A file path to the directory where the results will be saved. The default is the current working directory. The path can be relative or absolute.

verbose

A logical value specifying if output should be written to the screen or not. Good for testing and exploring your data but can be turned off when output indicates errors. The printing doesn't affect any of the returned objects, instead, printing makes it easier to see if the data has errors and what was done to rectify them in the given function. Sorry, but the default is to always print to the screen, i.e., verbose = TRUE, so you do not miss out on anything. This is standard practice for packages in the nwfsc-assess organization.

Value

An RData object is saved to the disk and the pulled data are returned as a data frame. The saved data can be read back in using load, but note that the object is named bds.pacfin inside of the .RData file, and thus, the object will retain this name within your work space. In February of 2021, the returned object was renamed from out

to bds.pacfin in an attempt to be more clear.

Upon download, the data are changed from a long table to a wide table using the combination of unique FISH_ID and AGE_SEQUENCE_NUMBER. Users are warned if verbose = TRUE when there are non-unique FISH_IDs and these data are automatically removed from the data frame. This change from long to wide allows for rows equating to a single fish with columns containing information about all measurements for that fish. Multiple age reads and information about those reads such as age reader will be in the columns. So, if your data had at most three duplicate reads, then you will see numbers up to 3 pasted onto the end of relevant column names. This is to help you match duplicate reads to other relevant information. Not all double reads are currently available within PacFIN and users should contact the ageing labs if they wish to inform ageing-error matrices.

AGE_COUNT is a somewhat cryptic column name and does not always make sense when compared to AGE_SEQUENCE_NUMBER. It was determined that the former is useful to identify how many potential agers were exposed to this fish. For example, if AGE_SEQUENCE_NUMBER has a maximum value of three for a given FISH_ID, then you can expect AGE_COUNT to be three for all three rows in the PACFIN database for that fish. This is not always true though. Sometimes, not all AGE_SEQUENCE_NUMBERs are present and they can skip numbers for a given FISH_ID, and in this case, AGE_COUNT will be the maximum AGE_SEQUENCE_NUMBER for a given FISH_ID.

FINAL_FISH_AGE_IN_YEARS is known as the best age for a given fish. This will not always match an age reader or be a number determinable from the individual age reads in AGE_IN_YEARS. Patrick explained to me that when age reads do not agree, particularly for younger fish, then the senior reader will work together with the junior reader to determine an agreed-upon age. Other times, the senior reader's value will always be used, or it could be that together they determine that they were both wrong and a new age is proposed as the resolved age. Nevertheless, it can be quite messy and there is no way to predict the best age.

See also

Author

John R. Wallace, Kelli F. Johnson

Examples

if (FALSE) {
# Enter password in place of xxx below
mypw <- "xxx"
pd <- PullBDS.PacFIN(pacfin_species_code = "POP", password = mypw)
}