Pull Catch Data from PacFIN Database

Pull Catch Data from the Comprehensive Fish Ticket table in the PacFIN database.

PullCatch.PacFIN(
  pacfin_species_code,
  username = getUserName("PacFIN"),
  password,
  savedir = getwd(),
  addnominal = TRUE,
  verbose = FALSE
)

Arguments

pacfin_species_code

A character string for the PacFIN species code that you desire to pull data for. Known also as "SPID" in the PacFIN database if using legacy sql scripts. An example for sablefish would be pacfin_species_code = "SABL". A list of pacfin_species_code can be found online: https://pacfin.psmfc.org/pacfin_pub/marketcat.php

username

The username you would like to use to access data from datasourcename. The default is to generate username from your computer name.

password

Your password for the datasourcename.

savedir

A file path to the directory where the results will be saved. The default is the current working directory. The path can be relative or absolute.

addnominal

A logical, the default TRUE adds the nominal PacFIN species code to the list of those searched for. FALSE leaves pacfin_species_code as input by the user which will miss catch for species such as dover sole that have more than one species code within PacFIN, e.g., pacfin_species_code = c("DOVR", "DVR1"). Users can also input a vector to pacfin_species_code if you want to specify the species and the nominal species code without searching for it. This is helpful for when you are getting data for two species and you only want nominal catch for one or if you only wanted nominal catch. Nominal species code will be searched for using PullNominal.PacFIN if the input value for addnominal is TRUE.

verbose

A logical value specifying if output should be written to the screen or not. Good for testing and exploring your data but can be turned off when output indicates errors. The printing doesn't affect any of the returned objects, instead, printing makes it easier to see if the data has errors and what was done to rectify them in the given function. Sorry, but the default is to always print to the screen, i.e., verbose = TRUE, so you do not miss out on anything. This is standard practice for packages in the nwfsc-assess organization.

Value

RData frames are saved to the disk and the pulled data are returned as a data frame.

CompFT - pulled data

Research.Tribal.Catch - summary by year, state, and fleet; fleets are as follows: limited entry (LE), open access (OA), treaty indian (TI), research (R), and unknown (XX)

Catch.INPFC - catch by INPFC area

Catch.PSMFC - catch by PSMFC area

Details

Species with special considerations: URCK: There is an rockfish (URCK) category that consists on unassigned rockfish landings. The majority of the catch is prior to 2001. Currently, there is no agreed upon methodology to parsed these landing out to specific rockfish species. At present, landings in this category are not included in species-specific rockfish catch pulls. POP: In PacFIN there are four species code that can be associated with Pacific ocean perch. These are POP, POP1, POP2, UPOP. The POP1 is general shelf/slope rockfish and not Pacific ocean perch specific landings. These records occur only in Oregon. As of the 2017 assessment, these records should be removed from the catch file.

See also

Author

Kelli F. Johnson