Pull the species information table and return nominal code

Pull the species information table with PACFIN_SPECIES_CODE and PACFIN_COMMON_NAME from PacFIN to determine the nominal code for a given species code.

PullNominal.PacFIN(pacfin_species_code, username, password)

Arguments

pacfin_species_code

A character string for the PacFIN species code that you desire to pull data for. Known also as "SPID" in the PacFIN database if using legacy sql scripts. An example for sablefish would be pacfin_species_code = "SABL".

username

The username you would like to use to access data from datasourcename. The default is to generate username from your computer name.

password

Your password for the datasourcename.

Value

A vector of character values, where each value is a PACFIN_SPECIES_CODE that pertains to the input pacfin_species_code. NA is returned if no values are found.