Filter samples for appropriate ages and agemethods.

cleanAges is now defunct and and users now only need to run cleanPacFIN() rather than both functions.

cleanAges(
  Pdata,
  keep_age_methods = c("B", "S", ""),
  minAge = 1,
  maxAge = NULL,
  CLEAN = TRUE
)

Arguments

Pdata

A PacFIN dataset

keep_age_methods

By default, methods "B","S" and "" are retained.

minAge

Minimum age to retain; default is 1.

maxAge

Maximum age to retain; default (NULL) returns all ages greater than minAge.

CLEAN

Default is TRUE. If FALSE, return the input data unchanged.

Value

The input data filtered for age methods and minimum and maximum ages specified, with added column age. A brief report on filtering is output to the console.

Details

Age methods "1" and "2" are equivalent to "B" and "S" respectively, and are reset to those values.

No ages less than one occur in the commercial dataset. Zero ages are equivalent to NA.

See also

Author

Andi Stephens