Combine CalCOM and PacFIN data

Adds required PacFIN columns to CalCOM data, initializing them to meaningful values as appropriate and translating from the CalCOM values when they exist in a different format.

combineCalCOM(Pdata, CalCOM)

Arguments

Pdata

Output from PullBDS.PacFIN().

CalCOM

A data frame supplied, more than likely be E.J. Dick or California Department of Fish and Wildlife containing biological samples for flatfish species prior to 1990.

Value

Returns a combined dataset in PacFIN format.

Details

This function was made defunct in March 2021 in commit d5c0355 because CalCOM data were expected to be available in PacFIN and restored in April 2023 after discovering that CalCOM data for flatfish prior to 1990 is not currently available in PacFIN, as discussed in issue #101 because the early data are "in separate tables because they are bin samples (# of fish) rather than cluster samples (weight based). PacFIN doesn't access those tables."

Coding of data from CalCOM to PacFIN is done for dates, genders, and areas. PSMFC_AREAs are derived from CalCOM "PORT"s.

Other data that are not reported in CalCOM datasets (e.g., SAMPLE_TYPE and SAMPLE_METHOD) are initialized to typical default values kept when cleaning data so that the CalCOM data aren't rejected by cleanPacFIN().

See also

Author

Melissa Haltuch, Andi Stephens, and Ian G. Taylor