Plot cleaned fishery age- or length-composition data.

Save several figures to your computer that summarize the composition samples that are available for the given species. None of the figures are meant to be included in the stock assessment document, but rather, are for exploring the data.

plotCleaned(Pdata, savedir = getwd())

Arguments

Pdata

A data frame of biological samples originating from the Pacific Fishieries Information Network (PacFIN) data warehouse, which originated in 2014. Data are pulled using sql calls, see PullBDS.PacFIN().

savedir

A file path to the directory where the results will be saved. The default is the current working directory. The path can be relative or absolute.

Value

The following figures are saved to savedir:

PacFIN_comp_Nbystate.png

PacFIN_comp_distributions.png

PacFIN_comp_NbyGRID.png

PacFIN_comp_depth.png

PacFIN_comp_geargroup.png

PacFIN_comp_INPFC.png

PacFIN_comp_lengthvage.png

See also

Author

Andi Stephens, Kelli F. Johnson