The number of samples per combination of ageing methods

Uses dplyr::count() to summarize the number of samples across the various ageing methods (i.e., AGE_METHOD[0-9]+) that were used in the passed data frame. Sometimes the ageing method will differ between double reads or between years.

summaryAgeMethod(Pdata, verbose = FALSE)

Arguments

Pdata

A data frame of biological samples originating from the Pacific Fishieries Information Network (PacFIN) data warehouse, which originated in 2014. Data are pulled using sql calls, see PullBDS.PacFIN().

verbose

A logical specifying if output should be written to the screen or not. Good for testing and exploring your data but can be turned off when output indicates information that you already know. The printing of output to the screen does not affect any of the returned objects. The default is to always print to the screen, i.e., verbose = TRUE.

Value

A data frame of counts (n) for each age method with the following columns:

AGE_METHOD[0-9]*: where the number at the end of the column indicates which age reader, e.g., first, second, third, etc., the method comes from. This column contains a letter or number indicating the ageing method used. There will be one column per number of reads available, i.e., if at least one fish was read by four reader, then there will be four columns starting with AGE_METHOD.

Age method for best age: provides data on all methods used across age readers, for example if both break and burn and surface read information is available for a given fish then this column will have B--S as an entry. This information is important because sometimes the best age is based off multiple methods and it is difficult to know which ageing error to assign the entry to.

n: the number of samples available for the given methods listed in that row.

If verbose, then text explaining the summary is also printed to the screen prior to returning the table.

Author

Kelli F. Johnson