Get age information for each fish

Get age information for each fish, or row of the data. This functionality was originally in cleanAges, moved to cleanPacFIN, and is now compartmentalized here, but called in cleanPacFIN and stored in the Age column.

getAge(Pdata, verbose = TRUE, keep, col.bestage = "FISH_AGE_YEARS_FINAL")

Arguments

Pdata

A data frame of biological samples originating from the Pacific Fishieries Information Network (PacFIN) data warehouse, which originated in 2014. Data are pulled using sql calls, see PullBDS.PacFIN().

verbose

A logical specifying if output should be written to the screen or not. Good for testing and exploring your data but can be turned off when output indicates information that you already know. The printing of output to the screen does not affect any of the returned objects. The default is to always print to the screen, i.e., verbose = TRUE.

keep

A vector of values that represent what you want to keep. Values of NA, "", and numeric values are acceptable. Often, it is helpful to run `unique(, useNA = "always") on the relevant data prior to running the function to see what types are present.

col.bestage

A character value noting the column name that contains the best age to use going forward for a given fish. Over time, this has been FISH_AGE_YEARS_FINAL and FINAL_FISH_AGE_IN_YEARS for PacFIN.

Value

A vector of ages in years with a length the same as the number of rows present in the input data frame Pdata. If all ages are NA, then a vector of NA values is quickly returned.

Details

Age types listed in keep (or derivatives of those) are used to filter for ages that should be kept, i.e., not returned as NA. If FISH_AGE_YEARS_FINAL doesn't have an entry, where this should be the best age available, then the age reads in age1, age2, ... are averaged over and rounded to the next integer (i.e., ceiling) if their methods are acceptable. When FISH_AGE_YEARS_FINAL does have an entry, this age is kept if one of the methods in AGE_METHOD is listed in keep. It is currently unclear how FISH_AGE_YEARS_FINAL is determined because it doesn't always equal the value from the first age reader. Sometimes, it is ceiling(mean()), other times it appears to default to break and burn, and other random combos. We use the best judgment of those who created the column that they know what they are doing, but please feel free to report errors in PacFIN to the github issues or the package maintainer after contacting the state representative.

The input argument of keep can contain a mix of numeric and character values, but all values will be converted to character values. See below for conversions, where it is easiest to just supply a vector of values shown on the left before the double equal sign because all values in your data set to the right will be converted prior to searching for those in keep. For example, keep = c("B", "L") will keep all "BB", 1, and 6 entries as well.

"B" == 1, "BB", or break and burn

"S" == 2, "SR" (California), or surface read

3 == scales (Oregon data)

"T" == 4, "TS" (California), "X" (Washington), or thin section

"O" == 5 or optical scanner

"L" == 6 or age derived from length

9 == unable to read or age

U == unknown

The final age assigned to a fish when there are multiple age reads depends on the ageing lab. See issue #49 for a semi-detailed discussion. Sometimes, reconciling happens where the best age will not be equal to any age read from a specific reader but the readers get together to determine the best read. Washington uses the following order of operations to assign the best age given there are multiple reads. Note, that they will never report more than three reads because that is the limitation of their database.

age1 if only one read;

randomly chosen age between age-read 1 and age-read 2 if the difference between the two reads is less than three;

if the difference between two reads is more than three, then the best age is manually keyed in from that designated by a human;

if three reads are equal, then the value is automatically assigned; and

if there are three reads and they are not equal then an age reader must choose the final value.