Aggregate composition data by length, Age, or age-at-length according to the given stratification.

Workflow

getComps is run subsequently to getExpansion_2.

getComps(
  Pdata,
  strat = NULL,
  Comps = c("AAL", "LEN", "AGE"),
  defaults = c("fleet", "fishyr", "season"),
  verbose = FALSE,
  ...
)

Arguments

Pdata

A data frame of biological samples originating from the Pacific Fishieries Information Network (PacFIN) data warehouse, which originated in 2014. Data are pulled using sql calls, see PullBDS.PacFIN().

strat

A vector of column names to stratify the output over. For example, if you wish to summarize by ageing method, then the argument would look like strat = "agemethod". Or if you want to look at fleets and gear, then use strat = c("fleet","usegear"). todo: deprecate defaults

Comps

The type of composition data to create. Options are length ("LEN", age ("AGE"), or conditional age-at-length ("AAL"). For some functions, a value of NA will allow for non-specific code to be run, i.e., not based on the composition type. Note that the entry needs to be in all upper-case letters.

defaults

The default stratification columns which will typically be left at their default value of c('fleet', 'fishyr', 'season').

verbose

A logical value specifying if output should be written to the screen or not. Good for testing and exploring your data but can be turned off when output indicates errors. The printing doesn't affect any of the returned objects, instead, printing makes it easier to see if the data has errors and what was done to rectify them in the given function. Sorry, but the default is to always print to the screen, i.e., verbose = TRUE, so you do not miss out on anything. This is standard practice for packages in the nwfsc-assess organization.

...

Pass additional arguments to getcomps_long, such as dropmissing = FALSE where the default behavior is dropmissing = TRUE

Value

A dataframe with composition data specific to the type specified in Comps for males, females, and unsexed records.

Details

The aggregation is of the Pdata$Final_Sample_Size column value, which should be set to the desired expansion:

Pdata$Final_Sample_Size = Pdata$Expansion_Factor_1

or

Pdata$Final_Sample_Size = Pdata$Expansion_Factor_1 * Pdata$Expansion_Factor_2

The default stratification is by fleet, fishyr, and season. Columns lengthcm, Age or both are added depending on the Comps argument.

Author

Andi Stephens, Kelli F. Johnson